+++ FeedModule/FeediOSTests/FeedViewControllerTests.swift
@@ -341,39 +341,5 @@ final class FeedViewControllerTests: XCTestCase {
     }
 }
 
-// MARK: - DSL Helper Extensions
-
-private extension FeedImageCell {
-    
-    var isShowingLocation: Bool {
-        return !locationContainer.isHidden
-    }
-    var locationText: String? {
-        return locationLabel.text
-    }
-    
-    var descriptionText: String? {
-        return descriptionLabel.text
-    }
-    
-    var isShowingImageLoadingIndicator: Bool {
-        return feedImageContainer.isShimmering
-    }
-    
-    var renderedImage: Data? {
-        return feedImageView.image?.pngData()
-    }
-    
-    var retryActionIsVisible: Bool {
-        return !retryButton.isHidden
-    }
-    
-    func simulateRetryAction() {
-        retryButton.simulateTap()
-    }
-    
-}
-
-
