+++ FeedModule/FeediOSTests/Helpers/FeedImageCell+Helpers.swift
@@ -1,0 +1,41 @@
+//
+//  FeedImageCell+Helpers.swift
+//  FeediOSTests
+//
+//  Created by Omran Khoja on 2/19/22.
+//
+
+import FeediOS
+import Foundation
+
+extension FeedImageCell {
+    
+    var isShowingLocation: Bool {
+        return !locationContainer.isHidden
+    }
+    var locationText: String? {
+        return locationLabel.text
+    }
+    
+    var descriptionText: String? {
+        return descriptionLabel.text
+    }
+    
+    var isShowingImageLoadingIndicator: Bool {
+        return feedImageContainer.isShimmering
+    }
+    
+    var renderedImage: Data? {
+        return feedImageView.image?.pngData()
+    }
+    
+    var retryActionIsVisible: Bool {
+        return !retryButton.isHidden
+    }
+    
+    func simulateRetryAction() {
+        retryButton.simulateTap()
+    }
+    
+}
+
